+++ can/firmware/platform_specific_hal_conf.h.in
