+++ can/firmware/utils.c
@@ -1,34 +1,28 @@
-#include "can/firmware/utils.hpp"
+#include "can/firmware/utils.h"
 
-#pragma GCC diagnostic push
-// NOLINTNEXTLINE(clang-diagnostic-unknown-warning-option)
-#pragma GCC diagnostic ignored "-Wvolatile"
-#include "platform_specific_hal_fdcan.h"
-#pragma GCC diagnostic pop
-
-using namespace can_bus;
+#include "platform_specific_hal_conf.h"
 
 /**
  * Convert a CanFilterConfig to a hal value
  * @param config can filter config enum class
  * @return hal defined constant
  */
-auto hal_can_utils::filter_config_to_hal(CanFilterConfig config) -> uint32_t {
-    auto result = FDCAN_FILTER_DISABLE;
+uint32_t filter_config_to_hal(enum CanFilterConfig config)  {
+    uint32_t result = FDCAN_FILTER_DISABLE;
     switch (config) {
-        case CanFilterConfig::reject:
+        case reject:
             result = FDCAN_FILTER_REJECT;
             break;
-        case CanFilterConfig::to_fifo0:
+        case to_fifo0:
             result = FDCAN_FILTER_TO_RXFIFO0;
             break;
-        case CanFilterConfig::to_fifo1:
+        case to_fifo1:
             result = FDCAN_FILTER_TO_RXFIFO1;
             break;
-        case CanFilterConfig::to_fifo0_high_priority:
+        case to_fifo0_high_priority:
             result = FDCAN_FILTER_TO_RXFIFO0_HP;
             break;
-        case CanFilterConfig::to_fifo1_high_priority:
+        case to_fifo1_high_priority:
             result = FDCAN_FILTER_TO_RXFIFO1_HP;
             break;
         default:
@@ -42,16 +36,16 @@ auto hal_can_utils::filter_config_to_hal(CanFilterConfig config) -> uint32_t {
  * @param type can filter type enum class
  * @return hal defined constant
  */
-auto hal_can_utils::filter_type_to_hal(CanFilterType type) -> uint32_t {
-    auto result = FDCAN_FILTER_RANGE;
+uint32_t filter_type_to_hal(enum CanFilterType type)  {
+    uint32_t result = FDCAN_FILTER_RANGE;
     switch (type) {
-        case CanFilterType::exact:
+        case exact:
             result = FDCAN_FILTER_DUAL;
             break;
-        case CanFilterType::range:
+        case range:
             result = FDCAN_FILTER_RANGE;
             break;
-        case CanFilterType::mask:
+        case mask:
             result = FDCAN_FILTER_MASK;
             break;
         default:
@@ -65,39 +59,39 @@ auto hal_can_utils::filter_type_to_hal(CanFilterType type) -> uint32_t {
  * @param length length
  * @return hal encoded length
  */
-auto hal_can_utils::length_to_hal(CanFDMessageLength length) -> uint32_t {
+uint32_t length_to_hal(enum CanFDMessageLength length)  {
     switch (length) {
-        case CanFDMessageLength::l0:
+        case l0:
             return FDCAN_DLC_BYTES_0;
-        case CanFDMessageLength::l1:
+        case l1:
             return FDCAN_DLC_BYTES_1;
-        case CanFDMessageLength::l2:
+        case l2:
             return FDCAN_DLC_BYTES_2;
-        case CanFDMessageLength::l3:
+        case l3:
             return FDCAN_DLC_BYTES_3;
-        case CanFDMessageLength::l4:
+        case l4:
             return FDCAN_DLC_BYTES_4;
-        case CanFDMessageLength::l5:
+        case l5:
             return FDCAN_DLC_BYTES_5;
-        case CanFDMessageLength::l6:
+        case l6:
             return FDCAN_DLC_BYTES_6;
-        case CanFDMessageLength::l7:
+        case l7:
             return FDCAN_DLC_BYTES_7;
-        case CanFDMessageLength::l8:
+        case l8:
             return FDCAN_DLC_BYTES_8;
-        case CanFDMessageLength::l12:
+        case l12:
             return FDCAN_DLC_BYTES_12;
-        case CanFDMessageLength::l16:
+        case l16:
             return FDCAN_DLC_BYTES_16;
-        case CanFDMessageLength::l20:
+        case l20:
             return FDCAN_DLC_BYTES_20;
-        case CanFDMessageLength::l24:
+        case l24:
             return FDCAN_DLC_BYTES_24;
-        case CanFDMessageLength::l32:
+        case l32:
             return FDCAN_DLC_BYTES_32;
-        case CanFDMessageLength::l48:
+        case l48:
             return FDCAN_DLC_BYTES_48;
-        case CanFDMessageLength::l64:
+        case l64:
             return FDCAN_DLC_BYTES_64;
         default:
             return FDCAN_DLC_BYTES_64;
@@ -109,40 +103,40 @@ auto hal_can_utils::length_to_hal(CanFDMessageLength length) -> uint32_t {
  * @param length hal encoded length
  * @return CanFDMessageLength
  */
-auto hal_can_utils::length_from_hal(uint32_t length) -> CanFDMessageLength {
+enum CanFDMessageLength length_from_hal(uint32_t length)  {
     switch (length) {
         case FDCAN_DLC_BYTES_0:
-            return CanFDMessageLength::l0;
+            return l0;
         case FDCAN_DLC_BYTES_1:
-            return CanFDMessageLength::l1;
+            return l1;
         case FDCAN_DLC_BYTES_2:
-            return CanFDMessageLength::l2;
+            return l2;
         case FDCAN_DLC_BYTES_3:
-            return CanFDMessageLength::l3;
+            return l3;
         case FDCAN_DLC_BYTES_4:
-            return CanFDMessageLength::l4;
+            return l4;
         case FDCAN_DLC_BYTES_5:
-            return CanFDMessageLength::l5;
+            return l5;
         case FDCAN_DLC_BYTES_6:
-            return CanFDMessageLength::l6;
+            return l6;
         case FDCAN_DLC_BYTES_7:
-            return CanFDMessageLength::l7;
+            return l7;
         case FDCAN_DLC_BYTES_8:
-            return CanFDMessageLength::l8;
+            return l8;
         case FDCAN_DLC_BYTES_12:
-            return CanFDMessageLength::l12;
+            return l12;
         case FDCAN_DLC_BYTES_16:
-            return CanFDMessageLength::l16;
+            return l16;
         case FDCAN_DLC_BYTES_20:
-            return CanFDMessageLength::l20;
+            return l20;
         case FDCAN_DLC_BYTES_24:
-            return CanFDMessageLength::l24;
+            return l24;
         case FDCAN_DLC_BYTES_32:
-            return CanFDMessageLength::l32;
+            return l32;
         case FDCAN_DLC_BYTES_48:
-            return CanFDMessageLength::l48;
+            return l48;
         case FDCAN_DLC_BYTES_64:
-            return CanFDMessageLength::l64;
+            return l64;
     }
-    return CanFDMessageLength::l64;
+    return l64;
 }
