+++ common/firmware/CMakeLists.txt
@@ -7,6 +7,5 @@ endif()
 add_subdirectory("uart")
 add_subdirectory("spi")
 add_subdirectory("motor_driver")
-add_subdirectory("can")
 add_subdirectory("i2c")
