+++ gantry/firmware/CMakeLists.txt
@@ -16,8 +16,8 @@ set(GANTRY_FW_LINTABLE_SRCS
         ${CMAKE_CURRENT_SOURCE_DIR}/can_task.cpp
         ${COMMON_EXECUTABLE_DIR}/spi/spi_comms.cpp
         ${CAN_FW_DIR}/hal_can_bus.cpp
-        ${CAN_FW_DIR}/hal_can_message_buffer.cpp
-        ${CAN_FW_DIR}/utils.cpp
+        ${CAN_FW_DIR}/utils.c
+        ${CAN_FW_DIR}/hal_can.c
         )
 
 set(GANTRY_FW_NON_LINTABLE_SRCS
@@ -28,7 +28,6 @@ set(GANTRY_FW_NON_LINTABLE_SRCS
         ${CMAKE_CURRENT_SOURCE_DIR}/tim7.c
         ${CMAKE_CURRENT_SOURCE_DIR}/motor_hardware.c
         ${COMMON_EXECUTABLE_DIR}/spi/spi.c
-        ${COMMON_EXECUTABLE_DIR}/can/can.c
         ${COMMON_EXECUTABLE_DIR}/errors/errors.c
         ${COMMON_EXECUTABLE_DIR}/motor_driver/motor.c
         )
