+++ gantry/firmware/can_task.cpp
@@ -1,5 +1,6 @@
 #include "common/firmware/can_task.hpp"
 
+#include "can/firmware/hal_can.h"
 #include "can/core/device_info.hpp"
 #include "can/core/dispatch.hpp"
 #include "can/core/freertos_can_dispatch.hpp"
@@ -9,6 +10,7 @@
 #include "can/core/message_handlers/move_group_executor.hpp"
 #include "can/core/message_writer.hpp"
 #include "can/core/messages.hpp"
+#include "can/firmware/hal_can_bus.hpp"
 #include "common/core/freertos_message_queue.hpp"
 #include "common/core/freertos_task.hpp"
 #include "common/firmware/errors.h"
@@ -21,11 +23,11 @@
 #pragma GCC diagnostic push
 // NOLINTNEXTLINE(clang-diagnostic-unknown-warning-option)
 #pragma GCC diagnostic ignored "-Wvolatile"
-#include "can/firmware/hal_can_bus.hpp"
-#include "can/firmware/hal_can_message_buffer.hpp"
-#include "common/firmware/can.h"
+//#include "can/firmware/hal_can_bus.hpp"
+//#include "can/firmware/hal_can_message_buffer.hpp"
+//#include "common/firmware/can.h"
 #include "motor_hardware.h"
-#include "platform_specific_hal_conf.h"
+//#include "platform_specific_hal_conf.h"
 #pragma GCC diagnostic pop
 
 using namespace freertos_task;
@@ -53,8 +55,7 @@ static constexpr NodeId node_from_axis(GantryAxisType which) {
 static auto my_axis_type = get_axis_type();
 static auto my_node_id = node_from_axis(my_axis_type);
 
-extern FDCAN_HandleTypeDef fdcan1;
-static auto can_bus_1 = HalCanBus(&fdcan1);
+static auto can_bus_1 = HalCanBus(can_get_device_handle());
 static auto message_writer_1 = MessageWriter(can_bus_1, my_node_id);
 
 static freertos_message_queue::FreeRTOSMessageQueue<Move> motor_queue(
@@ -157,10 +158,28 @@ static auto dispatcher = Dispatcher(
     motor_dispatch_target, motion_group_dispatch_target,
     motion_group_executor_dispatch_target, device_info_dispatch_target);
 
+/**
+ * The type of the message buffer populated by HAL ISR.
+ */
+static auto read_can_message_buffer = freertos_message_buffer::FreeRTOSMessageBuffer<1024>{};
+static auto read_can_message_buffer_writer =
+    can_message_buffer::CanMessageBufferWriter(read_can_message_buffer);
+
+/**
+ *
+ * @param identifier
+ * @param data
+ * @param length
+ */
+void callback(uint32_t identifier, const uint8_t* data, uint8_t length) {
+    read_can_message_buffer_writer.send_from_isr(identifier, data, data + length);
+}
+
+
 [[noreturn]] void task_entry() {
-    if (MX_FDCAN1_Init(&fdcan1) != HAL_OK) {
-        Error_Handler();
-    }
+    can_bus_1.set_incoming_message_callback(callback);
+    can_bus_1.setup_node_id_filter(axis_type::get_node_id());
+    can_start();
 
     if (initialize_spi(my_axis_type) != HAL_OK) {
         Error_Handler();
@@ -168,11 +187,7 @@ static auto dispatcher = Dispatcher(
 
     motor.driver.setup();
 
-    can_bus::setup_node_id_filter(can_bus_1, my_node_id);
-    can_bus_1.start();
-
-    auto poller = FreeRTOSCanBufferPoller(
-        hal_can_message_buffer::get_message_buffer(), dispatcher);
+    auto poller = FreeRTOSCanBufferPoller(read_can_message_buffer, dispatcher);
     poller();
 }
