+++ include/can/firmware/hal_can.h
@@ -1,0 +1,73 @@
+#pragma once
+
+#include "can/core/types.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif  // __cplusplus
+
+/**
+ * Receive message callback.
+ */
+typedef void(*can_message_callback)(uint32_t identifier, const uint8_t* data, uint8_t length);
+
+/**
+ * HAL can handle type.
+ */
+typedef void * HAL_CAN_HANDLE;
+
+/**
+ * Start CAN.
+ */
+void can_start();
+
+
+/**
+ * Get CAN device handle.
+ * @return a handle to a device.
+ */
+HAL_CAN_HANDLE can_get_device_handle();
+
+
+/**
+ * Register a callback function to process new messages.
+ *
+ * This is called from an ISR
+ *
+ * @param callback a callback function.
+ */
+void can_register_message_callback(can_message_callback callback);
+
+
+/**
+ * Send a can message.
+ *
+ * @param handle handle to can device
+ * @param arbitration_id the arbitration id
+ * @param buffer pointer to start of buffer
+ * @param buffer_length length of buffer as defined by HAL
+ */
+void can_send_message(HAL_CAN_HANDLE handle, uint32_t arbitration_id, uint8_t* buffer,
+                      enum CanFDMessageLength buffer_length);
+
+
+
+/**
+ * Add a filter for incoming messages.
+ *
+ * @param handle handle to can device
+ * @param index index of the filter
+ * @param type the filter type (range, mask, exact match).
+ * @param config what the filter should do
+ * @param val1 filter value
+ * @param val2 filter value
+ */
+void can_add_filter(HAL_CAN_HANDLE handle, uint32_t index, enum CanFilterType type, enum CanFilterConfig config,
+                    uint32_t val1, uint32_t val2);
+
+
+
+#ifdef __cplusplus
+}  // extern "C"
+#endif  // __cplusplus
+
