--- include/can/firmware/platform_specific_hal_fdcan.h.in
@@ -1,7 +1,0 @@
-#if (${STM32L5} == 1)
-    #include "stm32l5xx_hal_fdcan.h"
-#elif (${STM32G4} == 1)
-    #include "stm32g4xx_hal_fdcan.h"
-#else
-    #error "No architecture set"
-#endif
